Added Validation for OverrideChecklistHandler

+++ application-manager/src/handlers/application/deleteApplicationDocument.handler.go
@@ -3,16 +3,45 @@ package applicationHandlers
 import (
 	"application-manager/src/orchestrators"
 	"application-manager/src/store/types"
+	"errors"
 	"net/http"
+	"strings"
 
 	"github.com/labstack/echo/v4"
+	"go.mongodb.org/mongo-driver/bson/primitive"
 )
+
+func validateDeleteApplicationDocumentInput(appId string, docId string) (bool, error) {
+	if strings.TrimSpace(appId) == "" {
+		return false, errors.New("appId is missing or is not a string")
+	}
+
+	if _, err := primitive.ObjectIDFromHex(appId); err != nil {
+		return false, errors.New("appId is not in valid ObjectID format")
+	}
+
+	if strings.TrimSpace(docId) == "" {
+		return false, errors.New("docId is missing or is not a string")
+	}
+
+	if _, err := primitive.ObjectIDFromHex(docId); err != nil {
+		return false, errors.New("docId is not in valid ObjectID format")
+	}
+
+	return true, nil
+}
 
 func DeleteApplicationDocument(c echo.Context) error {
 	responseData := types.ResponseBody{}
 
 	appId := c.Param("appId")
 	docId := c.Param("docId")
+
+	if valid, err := validateDeleteApplicationDocumentInput(appId, docId); !valid {
+		responseData.Message = "Error deleting document"
+		responseData.Data = err.Error()
+		return c.JSON(http.StatusBadRequest, responseData)
+	}
 
 	if err := orchestrators.DeleteApplicationDocument(appId, docId); err != nil {
 		responseData.Message = "Error deleting document"
